# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIXTYSIX API",
    "version": "0.2.0",
    "description": "구현 전 계약. B0 로그인 2종/공개 모집 반영. 서버 시간 권위, unknown request fields 거부, 업무 성공 쓰기만 멱등 캐시. 세부 권장 정책은 BACKEND-B0-SPEC.md 참조. 인증 cookie 실제 이름은 T03 adapter에서 고정. 모든 보호/인증 응답 Cache-Control: no-store."
  },
  "servers": [
    {
      "url": "/",
      "description": "동일 origin /v1 프록시; 배포 미구현"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/v1/admin/cohorts": {
      "post": {
        "operationId": "createCohort",
        "summary": "운영자 모집 개설",
        "tags": [
          "Operations"
        ],
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cohort"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "시작일 04:00 KST 계산, 모집 시작 < 시작 시각, minParticipants <= capacity. 최소 10은 권장값이며 기본값으로 강제하지 않음. 최근 재인증+감사 기록. admin 권한·세션·최근 5분 재인증은 성공 응답 재조회 때도 확인한다. 일반 업무와 동일한 멱등 키 공간/계정별 한도를 사용한다. 성공 상태·감사·멱등 응답은 같은 transaction이며 취소는 내부 알림 outbox도 함께 저장한다.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          },
          {
            "$ref": "#/components/parameters/Idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "habitId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "policyVersion": {
                    "type": "integer",
                    "enum": [
                      1
                    ]
                  },
                  "generation": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "recruitmentOpensAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "capacity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 30
                  },
                  "minParticipants": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 30
                  }
                },
                "required": [
                  "habitId",
                  "policyVersion",
                  "generation",
                  "startDate",
                  "recruitmentOpensAt",
                  "capacity",
                  "minParticipants"
                ]
              }
            }
          }
        },
        "x-required-role": "admin"
      },
      "get": {
        "operationId": "listAdminCohorts",
        "summary": "운영자 모집 목록",
        "tags": [
          "Operations"
        ],
        "description": "admin 역할만 허용. 재인증 없이 조회 가능, 시작일·ID 내림차순 최대 50개와 nextCursor를 반환한다. 조회 시 필요한 시작 판단을 수행한다.",
        "x-required-role": "admin",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "운영 모집 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "maxItems": 50,
                      "items": {
                        "$ref": "#/components/schemas/Cohort"
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        }
      }
    },
    "/v1/admin/cohorts/{cohortId}/cancel": {
      "post": {
        "operationId": "cancelCohort",
        "summary": "운영자 코호트 취소",
        "tags": [
          "Operations"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cohort"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "description": "최근 재인증+감사+참여자 안내 outbox. 시작 전 모집 취소만 첫 범위. 시작 후 긴급 중단은 별도 정책/명령 필요. admin 권한·세션·최근 5분 재인증은 성공 응답 재조회 때도 확인한다. 일반 업무와 동일한 멱등 키 공간/계정별 한도를 사용한다. 성공 상태·감사·멱등 응답은 같은 transaction이며 취소는 내부 알림 outbox도 함께 저장한다.",
        "parameters": [
          {
            "in": "path",
            "name": "cohortId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/Csrf"
          },
          {
            "$ref": "#/components/parameters/Idempotency"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "x-required-role": "admin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-sixtysix.session",
        "description": "HTTPS: __Host-sixtysix.session, Path=/·HttpOnly·Secure·SameSite=Lax·Domain 없음. 로컬 HTTP만 sixtysix.session. DB에는 토큰 SHA-256만 저장, 개발 기본 절대 만료 7일."
      }
    },
    "parameters": {
      "Idempotency": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "같은 동작의 네트워크 재시도는 같은 UUID, 변경된 입력은 새 UUID. method/path/body를 요청 hash에 포함."
      },
      "Csrf": {
        "in": "header",
        "name": "X-CSRF-Token",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "GET /v1/auth/context 발급값. 브라우저/세션에 결합, Origin 검사와 함께 사용."
      },
...
    },
    "responses": {
      "400": {
        "description": "입력 또는 인증 challenge/state 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "401": {
        "description": "유효 세션 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "403": {
        "description": "운영 권한 부족 또는 REAUTH_REQUIRED",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "404": {
        "description": "없거나 접근 불가한 업무 리소스",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "409": {
        "description": "업무/멱등/연결 충돌",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "429": {
        "description": "속도 제한",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            },
            "description": "재시도까지 초"
          }
        }
      },
      "503": {
        "description": "일시 장애",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Cohort": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "habitId": {
            "type": "string",
            "format": "uuid"
          },
          "generation": {
            "type": "string",
            "minLength": 1
          },
          "policyVersion": {
            "type": "integer",
            "minimum": 1
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "durationDays": {
            "type": "integer",
            "enum": [
              66
            ]
          },
          "capacity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          },
          "minParticipants": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          },
          "participantCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 30
          },
          "recruitmentOpensAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "recruiting",
              "active",
              "ended",
              "cancelled"
            ]
          },
          "finalSubmissionAt": {
            "type": "string",
            "format": "date-time"
          },
          "canJoin": {
            "type": "boolean"
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "habitId",
          "generation",
          "policyVersion",
          "startsAt",
          "startDate",
          "durationDays",
          "capacity",
          "minParticipants",
          "participantCount",
          "recruitmentOpensAt",
          "status",
          "finalSubmissionAt",
          "canJoin",
          "cancellationReason"
        ]
      },
      "Available": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "targetDay": {
            "type": "integer",
            "minimum": 1,
            "maximum": 66
          },
          "targetDate": {
            "type": "string",
            "format": "date"
          },
          "late": {
            "type": "boolean"
          },
          "closesAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "ok",
          "targetDay",
          "targetDate",
          "late",
          "closesAt"
        ]
      },
      "Unavailable": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "BEFORE_START",
              "ALREADY_FILLED",
              "ENDED",
              "CANCELLED",
              "LEFT"
            ]
          }
        },
        "required": [
          "ok",
          "reason"
        ]
      },
      "Availability": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/Available"
          },
          {
            "$ref": "#/components/schemas/Unavailable"
          }
        ]
      },
...
      "ErrorDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "field": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_REQUEST",
                  "INVALID_TEXT",
                  "INVALID_MEDIA",
                  "INVALID_CHALLENGE",
                  "INVALID_OAUTH_STATE",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "REAUTH_REQUIRED",
                  "NOT_FOUND",
                  "COHORT_FULL",
                  "JOIN_CLOSED",
                  "ACTIVE_MEMBERSHIP_EXISTS",
                  "ALREADY_FILLED",
                  "PASS_EXHAUSTED",
                  "BEFORE_START",
                  "ENDED",
                  "TARGET_CHANGED",
                  "IDEMPOTENCY_CONFLICT",
                  "IDENTITY_ALREADY_LINKED",
                  "ACCOUNT_NOT_LINKED",
                  "LAST_IDENTITY",
                  "INVALID_STATE",
                  "RATE_LIMITED",
                  "TEMPORARILY_UNAVAILABLE",
                  "COHORT_EXISTS"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "requestId": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "$ref": "#/components/schemas/ErrorDetails"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}